feat(api): #60 [P1] 청약 매칭 DTO 검증 강화

+++ apps/api/src/announcement/dto/match-request.dto.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 
 export const matchRequestSchema = z.object({
   age: z.number().int().finite().min(0).max(150),
-  income: z.number().finite().min(0), // 만원 단위
+  income: z.number().int().finite().min(0), // 만원 단위
   homelessMonths: z.number().int().finite().min(0),
   dependents: z.number().int().finite().min(0).optional(),
   region: z.string().trim().min(1).optional(),
